86czyeauc: Create ImageType file for enum and interface
